[MCP] shopify: address review feedback (ShopifyClient)
